начало работы с апи регистрации
