+++ src/pages/authorization/authorization.service.ts
@@ -2,19 +2,24 @@ import AuthAPI from "../../api/auth-api";
 import Store from "../../modules/store";
 import constants from "../../constants";
 
-interface Data {
+export interface SignInData {
     login: string,
-    password: string
+    password: string,
 }
+
+const DEFAULT_PARAMS = {
+    loginInput: '',
+    passwordInput: ''
+};
 
 const api = new AuthAPI();
 const store = new Store();
 const currentPage = constants.routes.main;
 
-export const signIn = async (data: Data) => {
-    const response = await api.request(data);
+export const signIn = async (data: SignInData) => {
+    const response = await api.signIn(data);
     if (response?.status === 400) {
-        store.update({ loginInput: '', passwordInput: ''}, currentPage);
+        store.update(DEFAULT_PARAMS, currentPage);
     } else {
         // todo process values
     }
